requirements.txt
# Django Core
Django==5.2.4
asgiref==3.8.1
sqlparse==0.5.1

# REST Framework & Authentication
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
PyJWT==2.9.0

# Database
# psycopg2-binary==2.9.9
dj-database-url==2.1.0

# HTTP & API
requests==2.32.2
# urllib3 wird transitiv von requests/botocore installiert (botocore verlangt <2.1)
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7

# Caching & Redis (Production)
django-redis==5.4.0
redis==6.2.0
hiredis==2.3.2

# Security & Sanitization
bleach==6.1.0
django-csp==3.7

# Admin & UI
django-jazzmin==3.0.1
django-cors-headers==4.4.0
django-extensions==4.1

# Environment & Configuration
python-dotenv==1.1.0
python-decouple==3.8

# Production & Deployment
gunicorn==23.0.0
whitenoise==6.6.0

# Utilities
packaging==25.0
typing_extensions==4.12.2
tzdata==2025.2
python-docx==1.2.0

# AWS/Wasabi Integration
boto3==1.34.0
botocore==1.34.0

# Development & Testing (optional - uncomment if needed)
# pytest==7.4.3
# pytest-django==4.7.0
# factory-boy==3.3.0
# coverage==7.3.2
ruff==0.12.11

# Stripe Integration
dj-stripe==2.10.*